requirements.txt
ttkbootstrap
pyyaml
jinja2
